+++ src/test/java/roomescape/service/dto/time/ReservationTimeRequestTest.java
@@ -11,7 +11,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import roomescape.service.dto.reservation.ReservationTimeRequest;
 
 class ReservationTimeRequestTest {
